Sniper Monkey added
Create a Scriptable Object for the new Sniper Monkey
Create a Prefab for the Sniper Monkey using assets provided to you in the setup branch.
Use this Prefab in the Scriptable Object of the monkey
Add the Sniper Monkey Scriptable Object in the list of Monkey SOs being used in the game
Tweak the parameters of the sniper monkey in its Scriptable Object.